+++ zcash_client_backend/src/wallet.rs
@@ -65,11 +65,33 @@ impl NoteId {
 /// internal account ID and the pool to which funds were sent in the case of a wallet-internal
 /// output.
 #[derive(Debug, Clone)]
-pub enum Recipient {
+pub enum Recipient<N> {
     Transparent(TransparentAddress),
     Sapling(sapling::PaymentAddress),
     Unified(UnifiedAddress, PoolType),
-    InternalAccount(AccountId, PoolType),
+    InternalAccount(AccountId, N),
+}
+
+impl<N> Recipient<N> {
+    pub fn map_internal_account<B, F: FnOnce(N) -> B>(self, f: F) -> Recipient<B> {
+        match self {
+            Recipient::Transparent(t) => Recipient::Transparent(t),
+            Recipient::Sapling(s) => Recipient::Sapling(s),
+            Recipient::Unified(u, p) => Recipient::Unified(u, p),
+            Recipient::InternalAccount(a, n) => Recipient::InternalAccount(a, f(n)),
+        }
+    }
+}
+
+impl<N> Recipient<Option<N>> {
+    pub fn internal_account_transpose_option(self) -> Option<Recipient<N>> {
+        match self {
+            Recipient::Transparent(t) => Some(Recipient::Transparent(t)),
+            Recipient::Sapling(s) => Some(Recipient::Sapling(s)),
+            Recipient::Unified(u, p) => Some(Recipient::Unified(u, p)),
+            Recipient::InternalAccount(a, n) => n.map(|n0| Recipient::InternalAccount(a, n0)),
+        }
+    }
 }
 
 /// A subset of a [`Transaction`] relevant to wallets and light clients.
@@ -380,24 +402,41 @@ impl<NoteRef> orchard_fees::InputView<NoteRef> for ReceivedNote<NoteRef, orchard
 /// [ZIP 310]: https://zips.z.cash/zip-0310
 #[derive(Debug, Clone)]
 pub enum OvkPolicy {
-    /// Use the outgoing viewing key from the sender's [`ExtendedFullViewingKey`].
+    /// Use the outgoing viewing key from the sender's [`UnifiedFullViewingKey`].
     ///
     /// Transaction outputs will be decryptable by the sender, in addition to the
     /// recipients.
     ///
-    /// [`ExtendedFullViewingKey`]: sapling::zip32::ExtendedFullViewingKey
+    /// [`UnifiedFullViewingKey`]: zcash_keys::keys::UnifiedFullViewingKey
     Sender,
 
-    /// Use a custom outgoing viewing key. This might for instance be derived from a
-    /// separate seed than the wallet's spending keys.
+    /// Use custom outgoing viewing keys. These might for instance be derived from a
+    /// different seed than the wallet's spending keys.
     ///
     /// Transaction outputs will be decryptable by the recipients, and whoever controls
-    /// the provided outgoing viewing key.
-    Custom(sapling::keys::OutgoingViewingKey),
-
-    /// Use no outgoing viewing key. Transaction outputs will be decryptable by their
+    /// the provided outgoing viewing keys.
+    Custom {
+        sapling: sapling::keys::OutgoingViewingKey,
+        #[cfg(feature = "orchard")]
+        orchard: orchard::keys::OutgoingViewingKey,
+    },
+    /// Use no outgoing viewing keys. Transaction outputs will be decryptable by their
     /// recipients, but not by the sender.
     Discard,
+}
+
+impl OvkPolicy {
+    /// Constructs an [`OvkPolicy::Custom`] value from a single arbitrary 32-byte key.
+    ///
+    /// Outputs of transactions created with this OVK policy will be recoverable using
+    /// this key irrespective of the output pool.
+    pub fn custom_from_common_bytes(key: &[u8; 32]) -> Self {
+        OvkPolicy::Custom {
+            sapling: sapling::keys::OutgoingViewingKey(*key),
+            #[cfg(feature = "orchard")]
+            orchard: orchard::keys::OutgoingViewingKey::from(*key),
+        }
+    }
 }
 
 /// Metadata related to the ZIP 32 derivation of a transparent address.